+++ packages/core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.1.9 (2022-11-30)
 
 - Add experimental `whyframe:iframe` support
 - Support custom options for `createApp()`